+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lightstep-action-predeploy",
-  "version": "0.1.4",
+  "version": "0.1.5",
   "description": "GitHub Action to see service health, reported errors, and on-call information in pull requests.",
   "main": "index.js",
   "scripts": {
@@ -22,7 +22,7 @@
   },
   "devDependencies": {
     "@zeit/ncc": "^0.22.3",
-    "eslint": "^7.8.1",
-    "jest": "^26.4.2"
+    "eslint": "^7.12.1",
+    "jest": "^26.6.1"
   }
 }
